pyproject.toml
[tool.poetry]
name = "data_analysis_agent"
version = "0.1.0"
description = "Chat-based QA agent for parquet dataset"
authors = ["HK"]
license = "None"
readme = "README.md"
package-mode = true
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"

# Core scientific stack
numpy = "^1.26"
pandas = "^2.2"
pyarrow = "^11.0"      
scikit-learn = ">=1.3.2"
matplotlib = "^3.10.0" 
hdbscan = "^0.8.33"

# SQL / Parquet
duckdb = "^1.1.3"

## file
pyyaml = "^6.0"

# LLM & API clients
openai = "^1.40"
openai-agents ="^0.3.0"
anthropic = "^0.39"
tiktoken = "^0.7"
requests = "^2.32"
tqdm = "^4.66"

# LangChain ecosystem (for agent orchestration if needed)
langchain = "^0.3"
langchain-community = "^0.3"
langgraph = "^0.2"
langsmith = ">=0.3.45"

# Utilities
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"
pytest = "^8.0"
pytest-cov = "^5.0"
jupyter = ">=1.1.1"
ipykernel = ">=6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["agent"]

[tool.coverage.report]
show_missing = true
skip_covered = true